Add a header partial and change the default layout
Note that the header assumes that the user is logged in. Therefore, if you're
not logged in and you try to load the header, you will get a null pointer
exception. This is okay since we will only have one page (the login page) that
does not require the user to be logged in, and it will use a different layout.